fix win test
