+++ tests/git_utils/test_git_clone_if_repo_is_remote.py
@@ -7,7 +7,10 @@ from git import Repo
 
 from gto.git_utils import git_clone, git_clone_if_repo_is_remote
 from tests.git_utils.data import get_example_http_remote_repo
-from tests.skip_presets import skip_for_windows_py_lt_3_9
+from tests.skip_presets import (
+    only_for_windows_py_lt_3_8,
+    skip_for_windows_py_lt_3_9,
+)
 
 
 def test_if_repo_is_a_meaningless_string_then_leave_it_unchanged():
@@ -30,6 +33,16 @@ def test_if_repo_is_remote_url_then_clone_and_set_repo_to_its_local_path():
         mocked_git_clone.assert_called_once_with(
             repo=get_example_http_remote_repo(), dir=local_repo
         )
+
+
+@only_for_windows_py_lt_3_8
+def test_if_repo_is_remote_url_and_windows_os_error_then_hint_win_with_py_lt_3_9_may_be_the_cause():
+    with pytest.raises(OSError) as e:
+        decorated_func(repo=get_example_http_remote_repo(), spam=0, jam=3)
+    assert e.type in (NotADirectoryError, PermissionError)
+    assert "windows" in str(e)
+    assert "python" in str(e)
+    assert "< 3.9" in str(e)
 
 
 @git_clone_if_repo_is_remote
